slimserver-funplug rc1

+++ tarballs/squeezecenter/start-slimserver.sh
@@ -14,20 +14,27 @@ stop_cmd="slimserver_stop"
 
 slimserver_start()
 {
+	if pidof slimserver.pl >/dev/null 2>/dev/null; then
+		echo "Already running"
+	else
 
-	# create slimserver database if necessary
-	if [ ! -d /ffp/var/mysql/slimserver ]; then
-		/ffp/sbin/reset-slimserver-database.sh
+		# create slimserver database
+		if [ ! -d /ffp/var/mysql/slimserver ]; then
+			/ffp/sbin/reset-slimserver-database.sh
+		fi
+    
+		cd $(dirname $command)
+		# this su command uses busybox-specific syntax!
+		su nobody -p -c "nohup ./$(basename $command) $slimserver_flags >/dev/null 2>&1 &"
 	fi
-
-    cd $(dirname $command)
-    # this su command uses busybox-specific syntax!
-    su nobody -p -c "nohup ./$(basename $command) $slimserver_flags >/dev/null 2>&1 &"
 }
 
 slimserver_stop()
 {
-    kill $(cat /ffp/lib/squeezecenter/Logs/slimserver.pid)
+	if [ -r /ffp/lib/squeezecenter/Logs/slimserver.pid ]; then
+		kill $(cat /ffp/lib/squeezecenter/Logs/slimserver.pid)
+	fi
+	killall slimserver.pl 2>/dev/null
 }
 
 run_rc_command "$1"
